python/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yt-knowledge-ingest"
version = "0.1.0"
description = "Turn YouTube URLs into structured wiki-style Markdown using Gemini (multimodal) and customizable prompts"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "google-genai>=1.47.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"yt_knowledge_ingest" = ["prompts/readme.txt", "prompts/.gitignore"]

[project.scripts]
yt-knowledge-ingest = "yt_knowledge_ingest.cli:main"
